pyproject.toml
[project]
name = "dicom-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.4.1",
    "mcp[cli]>=1.3.0",
    "pydicom>=3.0.1",
    "pynetdicom>=2.0.2",
    "pyyaml>=6.0.1",
]
